Moved setting up session out of app.ts to its own file.
