+++ app/app.ts
@@ -5,53 +5,13 @@ import webpack from 'webpack';
 import webpackDevMiddleware from 'webpack-dev-middleware';
 import * as nunjucks from 'nunjucks';
 import webpackDevConfig from '../webpack/webpack.dev.js';
-
 import { router } from './routes';
-
-import expressSession from 'express-session';
-import * as redis from 'redis';
-import redisConnect from 'connect-redis';
-import config from 'config';
+import { setupSession } from './session';
 
 const port: number | string = process.env.PORT || 3000;
 const app: express.Application = express();
 
-const useRedis = config.get('session.useRedis') === 'true';
-const isSecure = config.get('session.cookie.secure') === 'true';
-
-if (useRedis) {
-  const redisStore: expressSession.Store = redisConnect(expressSession);
-  const redisClient = redis.createClient();
-
-  redisClient.on('error', (err) => {
-    // tslint:disable-next-line no-console
-    console.log('Redis error: ', err);
-  });
-  const store = new redisStore({
-    url: config.get('session.redis.url'),
-    ttl: config.get('session.redis.ttlInSeconds'),
-    client: redisClient
-  });
-  app.use(expressSession({
-    cookie: {
-      httpOnly: true,
-      maxAge: config.get('session.cookie.maxAgeInMs'),
-      secure: isSecure
-    },
-    resave: true,
-    saveUninitialized: true,
-    secret: config.get('session.redis.secret'),
-    rolling: true,
-    store
-  }));
-} else {
-  app.use(expressSession({
-    secret: config.get('session.redis.secret'),
-    resave: false,
-    saveUninitialized: true,
-    cookie: { secure: false }
-  }));
-}
+app.use(setupSession());
 
 nunjucks.configure([
   'views',
